Fix some possible value truncations
Casts from size_t to unsigned int, etc.

+++ src/Word.cpp
@@ -56,7 +56,7 @@ Word::Word(const Word& other) {
     }
 }
 
-unsigned int Word::getLength() {
+size_t Word::getLength() {
     return this->all->size();
 }
